requirements.txt
anyio==3.6.1
boto3==1.26.53
botocore==1.29.53
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
click==8.1.3
fastapi==0.80.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.1
msgpack==1.0.4
pydantic==1.9.2
pynvim==0.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.14
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
